+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `VueObserveVisibility` and `VueLazyload` dependency
 
+### Changed / Improved
+- Made navbar visible while scrolling(#622)
+- Added background blur effect(#622)
+
 ## [1.0.4] - 04.01.2020
 
 ### Added